Move Progress_bar_callback out of plugin
